+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - added icons to archive/unarchive and delete buttons in task actions panel
 - correctly placing unmaximized app window on startup
 - added support for defining supported OS in the extension and showing an info if the extension is not supported on the current OS
+- added support for using external libraries for extension UI components
 
 ## [0.67.0]